[CppLox][Debug] Refactored disassembler out of Chunk
Chunk had code to disassemble and print opcodes in a debug-friendly
format. Refactored that logic into a new Disassemble class that returns
a vector of VMInstr and a OpCodePrinter class that dumps instructions
onto stdout. This is to easy the creation of compiler tests that check
the correspondence between statements and emitted bytecode.