[CppLox][Object] Added ObjectType enum to differenciate ObjectPtr.
std::dynamic_pointer_cast it's not useful to differenciate between
ObjectPtr types, so it wasn't possible to know if we had a FunctionPtr
or a NativeFunctionPtr.
Added ObjectType enum to differenciate and also enable the execution of
native functions. Had to patch native functions because those don't have
a chunk of bytecode because they are only raw function pointers with
native code defined on C++.